+++ src/algorithm/solution/dohyoungK/baekjoon/stack/S2493.java
@@ -1,4 +1,4 @@
-package algorithm.solution.dohyoungK.stack;
+package algorithm.solution.dohyoungK.baekjoon.stack;
 
 import algorithm.problem.baejoon.stack.P2493;
